backend/requirements.txt
bottle==0.12.18
dotmap==1.3.13
requests==2.24.0
gevent==20.6.2
tinydb==4.1.1
python-telegram-bot
